Convert automations service to TypeScript
no ref

This chain should have no user impact.

All the types were already here in JSDoc, so it was easy to rewrite in
TS.
